Adding initial versions of ``Lambda C Compiler''.  Right now, things are starting to shape up to some degree...  That is, the compiler will actually parse some very basic C.  Error handling is [unfortunately] non-existent, no [real] assembler is produced, and the semantics are completely ignored in variables [that is, you can do:
	int main(void)
	{
		return sin(b);
	}

even though there is no `b' in the current scope.  Oh well.  On
the bright side, though, it is only a few days of effort.

And, progress appears to be relatively quick so far.  I'm also
trying to come up to speed on assembler---as I don't really know
any of that yet.  Alpha is going to be the first target that I
tackle, in all likelihood.  It may even be the last, depending of
how things go.  Heh.

Modules are literate, but have no documentation as yet.